pyproject.toml
[project]
name = "deal_data_extractor"
version = "0.1.0"
description = "Deal Data Extraction and Processing Application"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "mt5manager>=5.0.4874",
    "fastapi>=0.109.2",
    "uvicorn[standard]>=0.27.1",
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.3",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.9",
]
requires-python = ">=3.9"
readme = "README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
packages = ["src/deal_data_extractor"]

[project.scripts]
deal-data-extractor = "deal_data_extractor.main:app"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[tool.black]
line-length = 88
target-version = ['py39']
